Separate functor objects into one header file as they are not needed by the user.
